+++ test/caylir/graph/json_test.exs
@@ -5,7 +5,6 @@ defmodule Caylir.Graph.JSONLibraryTest do
     alias Caylir.Graph.JSONLibraryTest.JSONLibrary
 
     use Caylir.Graph,
-      otp_app: :caylir,
       config: [
         json_decoder: JSONLibrary,
         json_encoder: JSONLibrary
@@ -16,7 +15,6 @@ defmodule Caylir.Graph.JSONLibraryTest do
     alias Caylir.Graph.JSONLibraryTest.JSONLibrary
 
     use Caylir.Graph,
-      otp_app: :caylir,
       config: [
         json_decoder: {JSONLibrary, :decode_partial},
         json_encoder: {JSONLibrary, :encode_partial}
@@ -27,7 +25,6 @@ defmodule Caylir.Graph.JSONLibraryTest do
     alias Caylir.Graph.JSONLibraryTest.JSONLibrary
 
     use Caylir.Graph,
-      otp_app: :caylir,
       config: [
         json_decoder: {JSONLibrary, :decode_full, [[keys: :atoms]]},
         json_encoder: {JSONLibrary, :encode_full, []}